+++ packages/components/Callout/macos/FRNCallout.h
@@ -1,0 +1,14 @@
+//
+//  FRNCallout.h
+//  Pods
+//
+//  Created by Patrick Boyd on 9/24/24.
+//
+// Created as a dummy file to enable importing FRNCallout-Swift.h and invoking Swift functions from Obj-C
+//
+
+#ifndef FRNCallout_h
+#define FRNCallout_h
+
+
+#endif /* FRNCallout_h */