+++ packages/components/Callout/src/Callout.types.ts
@@ -2,9 +2,10 @@ import type * as React from 'react';
 import type { KeyboardMetrics, ViewStyle } from 'react-native';
 
 import type { IViewProps } from '@fluentui-react-native/adapters';
-import type { IFocusable } from '@fluentui-react-native/interactive-hooks';
 import type { IBackgroundColorTokens, IBorderTokens } from '@fluentui-react-native/tokens';
 import type { IRenderData } from '@uifabricshared/foundation-composable';
+
+import type { CalloutNativeCommands } from './CalloutNativeCommands.types';
 export const calloutName = 'Callout';
 
 /**
@@ -113,7 +114,7 @@ export interface ICalloutProps extends IViewProps, ICalloutTokens {
   /**
    * A RefObject to access the IFocusable interface. Use this to access the public methods and properties of the component.
    */
-  componentRef?: React.RefObject<IFocusable>;
+  componentRef?: React.RefObject<CalloutNativeCommands>;
 
   /**
    * Defines event redirection behaviors for pointer events relative to the Callout control.