+++ packages/components/Text/src/__tests__/Text.test.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 
-import { checkRenderConsistency, checkReRender } from '@fluentui-react-native/test-tools';
 import * as renderer from 'react-test-renderer';
 
 import { Text } from '../Text';
@@ -24,36 +23,6 @@ describe('Text component tests', () => {
     });
     const tree = renderer.create(<BoldText>All tokens</BoldText>).toJSON();
     expect(tree).toMatchSnapshot();
-  });
-
-  it('Text simple rendering does not invalidate styling', () => {
-    checkRenderConsistency(() => <Text>Default</Text>);
-    checkRenderConsistency(() => <Text variant="headerStandard">Default</Text>);
-  });
-
-  it('Text re-renders correctly', () => {
-    checkReRender(() => <Text>Default</Text>);
-    checkReRender(() => <Text variant="headerStandard">Default</Text>);
-  });
-
-  it('Text shares produced styles across multiple renders', () => {
-    const style = { color: 'black' };
-    checkRenderConsistency(() => <Text style={style}>Default</Text>);
-    checkRenderConsistency(() => (
-      <Text style={style} variant="headerStandard">
-        Default
-      </Text>
-    ));
-  });
-
-  it('Text re-renders correctly with style', () => {
-    const style = { color: 'blue' };
-    checkReRender(() => <Text style={style}>Default</Text>);
-    checkReRender(() => (
-      <Text style={style} variant="headerStandard">
-        Default
-      </Text>
-    ));
   });
 
   it('Text variants render correctly with style', () => {