+++ packages/experimental/Tooltip/src/__tests__/Tooltip.test.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 
 import { ButtonV1 } from '@fluentui-react-native/button';
-import { checkRenderConsistency, checkReRender } from '@fluentui-react-native/test-tools';
 import * as renderer from 'react-test-renderer';
 
 import { Tooltip } from '../Tooltip';
@@ -17,28 +16,4 @@ describe('Tooltip component tests', () => {
       .toJSON();
     expect(tree).toMatchSnapshot();
   });
-
-  it('Tooltip simple rendering does not invalidate styling', () => {
-    checkRenderConsistency(
-      () => (
-        <Tooltip content="Default Tooltip">
-          <ButtonV1>Default Tooltip</ButtonV1>
-        </Tooltip>
-      ),
-      2,
-    );
-  });
-
-  it('Tooltip re-renders correctly', () => {
-    checkReRender(
-      () => (
-        <Tooltip content="Render twice">
-          <ButtonV1>Render twice</ButtonV1>
-        </Tooltip>
-      ),
-      2,
-    );
-  });
-
-  // Feel free to add more tests here
 });