+++ packages/components/Menu/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/menu
 
-This log was last generated on Wed, 10 Jul 2024 19:11:52 GMT and should not be manually modified.
+<!-- This log was last generated on Wed, 05 Mar 2025 01:28:08 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 1.14.20
+
+Wed, 05 Mar 2025 01:28:08 GMT
+
+### Patches
+
+- Fix Menu tab navigation on win32 (krsiler@microsoft.com)
+
+## 1.14.19
+
+Wed, 29 Jan 2025 23:43:53 GMT
+
+### Patches
+
+- remove enzyme tests (sanajmi@microsoft.com)
 
 ## 1.14.6