+++ packages/components/TabList/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/tablist
 
-This log was last generated on Fri, 12 Apr 2024 21:02:38 GMT and should not be manually modified.
+<!-- This log was last generated on Wed, 29 Jan 2025 23:43:54 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.7.2
+
+Wed, 29 Jan 2025 23:43:54 GMT
+
+### Patches
+
+- remove enzyme tests (sanajmi@microsoft.com)
+
+## 0.7.0
+
+Tue, 28 Jan 2025 23:24:31 GMT
+
+### Minor changes
+
+- Add native "Ctrl + Tab" keyboard shortcut for TabList component. (winlarry@microsoft.com)
+
+## 0.6.12
+
+Thu, 16 Jan 2025 19:12:11 GMT
+
+### Patches
+
+- Make tablist test not async so that it runs faster (ruaraki@microsoft.com)
 
 ## 0.6.1